+++ modules/app-android/src/androidMain/kotlin/org/archivekeep/app/android/AndroidEnvironmentPaths.kt
@@ -7,7 +7,7 @@ class AndroidEnvironmentPaths(
 ) {
     fun getRegistryDatastoreFile() = dataDir.resolve("registry.preferences_pb")
 
-    fun getWalletDatastoreFile() = dataDir.resolve("credentials.jwe")
+    fun getWalletDatastoreFile() = dataDir.resolve("credentials.enc")
 
     fun getRepositoryMetadataMemoryDatastoreFile() = dataDir.resolve("repository_metadata_memory.preferences_pb")