+++ modules/app-core/src/commonMain/kotlin/org/archivekeep/app/core/persistence/credentials/PasswordProtectedDataStore.kt
@@ -1,0 +1,7 @@
+package org.archivekeep.app.core.persistence.credentials
+
+interface PasswordProtectedDataStore<T> : ProtectedDataStore<T> {
+    suspend fun create(password: String)
+
+    suspend fun unlock(password: String)
+}