+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/ui/components/feature/AppBar.kt
@@ -23,12 +23,13 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import org.archivekeep.app.core.persistence.credentials.PasswordProtectedDataStore
 import org.archivekeep.app.ui.components.base.interactivity.DraggableAreaIfWindowPresent
 import org.archivekeep.app.ui.components.designsystem.navigation.AppBarIconButton
 import org.archivekeep.app.ui.components.designsystem.theme.CColors
 import org.archivekeep.app.ui.components.feature.buttons.drawerIconButton
 import org.archivekeep.app.ui.domain.data.canUnlock
-import org.archivekeep.app.ui.domain.wiring.LocalWalletDataStore
+import org.archivekeep.app.ui.domain.wiring.LocalApplicationServices
 import org.archivekeep.app.ui.domain.wiring.LocalWalletOperationLaunchers
 
 @Composable
@@ -37,7 +38,7 @@ fun AppBar(
     onCloseRequest: (() -> Unit)?,
     content: @Composable RowScope.() -> Unit,
 ) {
-    val credentialStorage = LocalWalletDataStore.current
+    val credentialStorage = LocalApplicationServices.current.environment.walletDataStore as? PasswordProtectedDataStore
 
     val canUnlock = credentialStorage.canUnlock()