+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/ui/components/feature/errors/RepositoryLockedError.kt
@@ -14,10 +14,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import compose.icons.TablerIcons
 import compose.icons.tablericons.Lock
+import org.archivekeep.app.core.persistence.credentials.PasswordProtectedDataStore
 import org.archivekeep.app.core.utils.exceptions.RepositoryLockedException
 import org.archivekeep.app.ui.domain.data.canUnlock
+import org.archivekeep.app.ui.domain.wiring.LocalApplicationServices
 import org.archivekeep.app.ui.domain.wiring.LocalArchiveOperationLaunchers
-import org.archivekeep.app.ui.domain.wiring.LocalWalletDataStore
 import org.archivekeep.app.ui.domain.wiring.LocalWalletOperationLaunchers
 
 @Composable
@@ -27,7 +28,7 @@ fun RepositoryLockedError(
 ) {
     val repoOperationLaunchers = LocalArchiveOperationLaunchers.current
     val walletOperationLaunchers = LocalWalletOperationLaunchers.current
-    val credentialStorage = LocalWalletDataStore.current
+    val credentialStorage = LocalApplicationServices.current.environment.walletDataStore as? PasswordProtectedDataStore
 
     Text("Repository ${cause.uri} is locked")