+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/ui/dialogs/wallet/CreateWalletDialog.kt
@@ -12,12 +12,12 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.text.buildAnnotatedString
 import kotlinx.coroutines.CoroutineScope
 import org.archivekeep.app.core.persistence.credentials.Credentials
-import org.archivekeep.app.core.persistence.credentials.JoseStorage
+import org.archivekeep.app.core.persistence.credentials.PasswordProtectedDataStore
 import org.archivekeep.app.ui.components.designsystem.input.PasswordField
 import org.archivekeep.app.ui.components.feature.dialogs.SimpleActionDialogControlButtons
 import org.archivekeep.app.ui.components.feature.dialogs.operations.LaunchableExecutionErrorIfPresent
 import org.archivekeep.app.ui.dialogs.AbstractDialog
-import org.archivekeep.app.ui.domain.wiring.LocalWalletDataStore
+import org.archivekeep.app.ui.domain.wiring.LocalApplicationServices
 import org.archivekeep.app.ui.utils.Launchable
 import org.archivekeep.app.ui.utils.asAction
 import org.archivekeep.app.ui.utils.simpleLaunchable
@@ -26,12 +26,12 @@ import org.archivekeep.utils.loading.Loadable
 class CreateWalletDialog : AbstractDialog<CreateWalletDialog.State, CreateWalletDialog.VM>() {
     class VM(
         scope: CoroutineScope,
-        val joseStorage: JoseStorage<Credentials>,
+        val passwordProtectedDataStore: PasswordProtectedDataStore<Credentials>,
         val _onClose: () -> Unit,
     ) : IVM {
         val launchable =
             simpleLaunchable(scope) { password: String ->
-                joseStorage.create(password)
+                passwordProtectedDataStore.create(password)
                 onClose()
             }
 
@@ -70,8 +70,9 @@ class CreateWalletDialog : AbstractDialog<CreateWalletDialog.State, CreateWallet
         scope: CoroutineScope,
         onClose: () -> Unit,
     ): VM {
-        val joseStorage = LocalWalletDataStore.current
-        return remember { VM(scope, joseStorage, onClose) }
+        val applicationServices = LocalApplicationServices.current
+
+        return remember { VM(scope, applicationServices.environment.walletDataStore as PasswordProtectedDataStore, onClose) }
     }
 
     @Composable