+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/ui/domain/data/credentialsUtils.kt
@@ -2,14 +2,21 @@ package org.archivekeep.app.ui.domain.data
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import org.archivekeep.app.core.persistence.credentials.Credentials
-import org.archivekeep.app.core.persistence.credentials.JoseStorage
+import org.archivekeep.app.core.persistence.credentials.PasswordProtectedDataStore
+import org.archivekeep.app.core.persistence.credentials.PasswordProtectedJoseStorage
+import org.archivekeep.app.core.utils.ProtectedLoadableResource
 
-fun JoseStorage<Credentials>.canUnlockFlow() = autoloadFlow.map { it is JoseStorage.State.Locked }
+fun(PasswordProtectedDataStore<Credentials>?).canUnlockFlow() =
+    (this as? PasswordProtectedJoseStorage)?.data?.map {
+        it is ProtectedLoadableResource.PendingAuthentication
+    } ?: flowOf(false)
 
 @Composable
-fun JoseStorage<Credentials>.canUnlock() =
-    autoloadFlow
-        .collectAsState()
-        .value is JoseStorage.State.Locked
+fun (PasswordProtectedDataStore<Credentials>?).canUnlock() =
+    (this as? PasswordProtectedJoseStorage)
+        ?.autoloadFlow
+        ?.collectAsState()
+        ?.value is PasswordProtectedJoseStorage.State.Locked