+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/ui/domain/wiring/CompositionLocals.kt
@@ -5,8 +5,6 @@ import org.archivekeep.app.core.domain.archives.ArchiveService
 import org.archivekeep.app.core.domain.repositories.RepositoryService
 import org.archivekeep.app.core.domain.storages.StorageRegistry
 import org.archivekeep.app.core.domain.storages.StorageService
-import org.archivekeep.app.core.persistence.credentials.Credentials
-import org.archivekeep.app.core.persistence.credentials.JoseStorage
 import org.archivekeep.app.core.persistence.drivers.filesystem.FileStores
 import org.archivekeep.app.core.persistence.registry.RegistryDataStore
 import org.archivekeep.app.core.procedures.add.IndexUpdateProcedureSupervisorService
@@ -14,11 +12,11 @@ import org.archivekeep.app.core.procedures.addpush.AddAndPushProcedureService
 import org.archivekeep.app.core.procedures.sync.RepoToRepoSyncService
 import org.archivekeep.app.ui.domain.services.RepositoryOpenService
 
+val LocalApplicationServices = staticCompositionLocalOfNotProvided<ApplicationServices>()
+
 val LocalStorageService = staticCompositionLocalOfNotProvided<StorageService>()
 val LocalRepoService = staticCompositionLocalOfNotProvided<RepositoryService>()
 val LocalArchiveService = staticCompositionLocalOfNotProvided<ArchiveService>()
-
-val LocalWalletDataStore = staticCompositionLocalOfNotProvided<JoseStorage<Credentials>>()
 
 val LocalRepoToRepoSyncService = staticCompositionLocalOfNotProvided<RepoToRepoSyncService>()
 val LocalAddPushService = staticCompositionLocalOfNotProvided<AddAndPushProcedureService>()